requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0

# Video processing
yt-dlp>=2024.0.0
opencv-python>=4.9.0.0
imagehash>=4.3.1
Pillow>=10.4.0

# Audio/Transcription
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
sentencepiece>=0.1.99

# NLP for Bengali
sentence-transformers>=2.2.2
scikit-learn>=1.3.2

# Vision/Embeddings
ftfy>=6.1.3
regex>=2023.12.25

# API clients
google-generativeai>=0.3.0
requests>=2.31.0

# Utilities
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0

# Note: Aeneas needs to be installed separately
# See: https://github.com/readbeyond/aeneas
